backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.46
pydantic==2.12.5
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.10.1
reportlab==4.2.2
jinja2==3.1.4
alembic==1.13.2
psycopg2-binary==2.9.9
python-json-logger==2.0.7
sentry-sdk==2.8.0
pytest==8.3.2
httpx==0.27.0
